ClickBehavior code minor changed

+++ Src/FM79979Engine/Core/Common/StringToStructure.cpp
@@ -37,10 +37,10 @@ namespace FATMING_CORE
 	float	GetFloat(std::string e_str){ return GetFloat(e_str.c_str()); }
 	float	GetFloat(std::wstring e_str){ return GetFloat(e_str.c_str()); }
 
-	float	GetDouble(const char* e_str) { return atof(e_str); }
-	float	GetDouble(const wchar_t* e_str) { return GetDouble(UT::WcharToChar(e_str).c_str()); }
-	float	GetDouble(std::string e_str) { return GetDouble(e_str.c_str()); }
-	float	GetDouble(std::wstring e_str) { return GetDouble(e_str.c_str()); }
+	double	GetDouble(const char* e_str) { return atof(e_str); }
+	double	GetDouble(const wchar_t* e_str) { return GetDouble(UT::WcharToChar(e_str).c_str()); }
+	double	GetDouble(std::string e_str) { return GetDouble(e_str.c_str()); }
+	double	GetDouble(std::wstring e_str) { return GetDouble(e_str.c_str()); }
 
 	bool	GetBoolean(const char* e_str){ return atoi(e_str)>0?true:false; }
 	bool	GetBoolean(const wchar_t* e_str){ return GetBoolean(UT::WcharToChar(e_str).c_str()); }
